requirements.txt
sentence-transformers
numpy
PyMuPDF  # imported as fitz
Pillow
torch
transformers
requests
msgpack
gradio
reportlab
google-generativeai
